List helper task
The code defines a 'ListHelper' class with two class methods: greatest_frequency and doubles. 

The 'greatest_frequency' method counts the frequency of each element in the list (my_list).

The 'doubles' method iterates and filters out the items in the list (my_list) that have a count greater than or equal to 2.