Create Item suitcase hold
The code  defines 3 classes: 'Item', 'Suitcase', and 'CargoHold'

The class 'Item' represents an item with name and weight attributes

The Suitcase class represents a suitcase that holds items.
It has methods to add items, calculate the total weight and retrieve the heaviest item. As well as the 'print_items' method that prints all of the items and their weight in the suitcase

The 'CargoHold' holds multiple suitcases. It has methods to add suitcases and calculate the total weight of the cargo hold. The 'print_items' method prints the name and weight of each item in all the suitcases in the cargo hold